requirements.txt
Django==3.0.7
djangorestframework
psycopg2
django-crispy-forms
django-allauth
pillow
stripe
django-debug-toolbar
whitenoise
gunicorn
dj-database-url
django-autoslug
django-author
django-cors-headers
coreapi
pyyaml
